Moving WiFi functionallity out of main source file
