+++ ESP32LapTimer/src/ESP32LapTimer.cpp
@@ -46,25 +46,6 @@ void IRAM_ATTR adc_task(void* args) {
   }
 }
 
-void WiFiEvent(WiFiEvent_t event) {
-  switch(event) {
-    case SYSTEM_EVENT_AP_START:
-      log_i("Setting system hostname");
-      WiFi.softAPsetHostname("chorus32");
-      break;
-    case SYSTEM_EVENT_STA_START:
-      log_i("Setting system hostname");
-      WiFi.setHostname("chorus32");
-      break;
-    case SYSTEM_EVENT_STA_DISCONNECTED:
-      log_i("WiFi network disconnected, retrying...");
-      WiFi.begin(WIFI_SSID, WIFI_PASSWORD);
-    default:
-      break;
-  }
-}
-
-
 void setup() {
 
 #ifdef OLED
@@ -87,17 +68,7 @@ void setup() {
   //PowerDownAll(); // Powers down all RX5808's
   delay(250);
 
-  WiFi.onEvent(WiFiEvent);
-
-#if defined(WIFI_MODE_ACCESSPOINT)
-  InitWifiAP();
-#elif defined(WIFI_MODE_CLIENT)
-  if(!InitWifiClient()) {
-    log_i("Failed to connect to WiFi Network");
-    log_i("Starting up in AP mode instead!");
-    InitWifiAP();
-  }
-#endif
+  InitWifi();
 
   InitWebServer();
