+++ ESP32LapTimer/src/Wireless.h
@@ -1,6 +1,8 @@
 #ifndef __WIRELESS_H__
 #define __WIRELESS_H__
 
+#include <WiFi.h>
+
 void InitWifiAP();
 void handleDNSRequests();
 void dnsHandleRequests();
@@ -9,5 +11,7 @@ void airplaneModeOff();
 void toggleAirplaneMode();
 bool isAirplaneModeOn();
 bool InitWifiClient();
+void InitWifi();
+void WiFiEvent(WiFiEvent_t event);
 
 #endif // __WIRELESS_H__
